requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
exceptiongroup==1.2.2
fastapi==0.115.12
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
kagglehub @ file:///C:/b/abs_08f8j6y1m6/croot/kagglehub_1720558573125/work
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-multipart==0.0.20
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==14.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
uvicorn==0.34.2
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.17.2
